database_code/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
openai==0.28.0 